requirements.txt
google-adk
google-genai
fastapi
uvicorn[standard]
pydantic